# withauth/docs/swagger.yaml
basePath: /api/v1
definitions:
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  model.Album:
    properties:
      artist:
        type: string
      id:
        type: string
      price:
        type: number
      title:
        type: string
    type: object
  model.User:
    properties:
      password:
        type: string
      userName:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: vitor.marcos@gmail.com
    name: API Support - VMRC
    url: http://www.swagger.io/support
  description: This is a sample server albums.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API - Albums
  version: "1.0"
paths:
  /albums:
    get:
      consumes:
      - application/json
      description: Get Albums responds with the list of all albums as JSON.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Album'
            type: array
      summary: get albums list
      tags:
      - albums
    post:
      consumes:
      - application/json
      description: adds an album from JSON received in the request body.
      parameters:
      - description: Add Album
        in: body
        name: album
        required: true
        schema:
          $ref: '#/definitions/model.Album'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Album'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: post an album
      tags:
      - albums
  /albums/{id}:
    delete:
      consumes:
      - application/json
      description: Locates the album whose ID value matches the id and remove from
        collection.
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: delete an albums by id
      tags:
      - albums
    get:
      consumes:
      - application/json
      description: Get the album as JSON whose ID value matches the id.
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Album'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: get an albums by id
      tags:
      - albums
  /signin:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: logon the user.
      parameters:
      - in: formData
        name: password
        type: string
      - in: formData
        name: userName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: post an user login
      tags:
      - signin
  /signout:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: logout the user.
      parameters:
      - in: formData
        name: password
        type: string
      - in: formData
        name: userName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: post an user logout
      tags:
      - signout
  /users:
    get:
      consumes:
      - application/json
      description: Get users responds with the list of all users as JSON.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
      security:
      - ApiKeyAuth: []
      summary: get users list
      tags:
      - users
    post:
      consumes:
      - application/json
      description: adds an user from JSON received in the request body.
      parameters:
      - description: Add User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: post an user
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"